IMplement auth using policy

+++ app/Http/Controllers/QuestionsController.php
@@ -9,6 +9,10 @@ use Illuminate\Support\Facades\DB;
 
 class QuestionsController extends Controller
 {
+
+    public function __construct(){
+        $this->middleware('auth')->except('index', 'show');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -86,6 +90,7 @@ class QuestionsController extends Controller
     public function update(AskQuestionRequest $request, Question $question)
     {
         //
+        $this->authorize("update", $question);
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', 'Question Updated!!');
 
@@ -99,6 +104,7 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize("update", $question);
         $question->delete();
         return redirect('/questions')->with('success', "Delete Successful!!");
     }
